requirements.txt
Flask
Flask-Cors
libsql-client
python-dotenv
